Moved glue.qt.widgets.terminal to glue.app.qt.terminal
