+++ glue/app/qt/terminal.py
